x86/sev: Fix potential sign extension in amd_enc_status_change_finish()
drivers inclusion
category: bugfix
bugzilla: https://atomgit.com/openeuler/kernel/issues/9200
CVE: NA

(cherry-pick from OLK-6.6 commit 61b5e7fb9c9d)

--------------------------------------------------------------------------

When npages (int) is left-shifted by PAGE_SHIFT to calculate the size
in bytes, the operation is performed on a signed integer. If npages is
large enough, this can cause:

  1. Signed integer overflow (undefined behavior per C standard)
  2. Sign extension when the result is promoted to unsigned long for
     the enc_dec_hypercall() parameter

For example, In CSV3 guest or SEV-SNP guest, invoke set_memory_encrypted()
or set_memory_encrypted() with param numpages equal or larger than 0x80000
can reproduce the issue. In function amd_enc_status_change_finish(),
left shifting int value 0x80000 12 bits to unsigned long, the actual value
becomes 0xffffffff80000000, which is unexpected.

Cast npages to unsigned long before the shift to ensure:
  - The shift operation uses unsigned arithmetic
  - No sign extension occurs during parameter promotion
  - Correct size calculation for large page counts (>2^19 pages on 32-bit)

Fixes: ac3f9c9f1b37 ("x86/sev: Make enc_dec_hypercall() accept a size instead of npages")

+++ arch/x86/mm/mem_encrypt_amd.c
@@ -364,7 +364,7 @@ static bool amd_enc_status_change_finish(unsigned long vaddr, int npages, bool e
 		snp_set_memory_private(vaddr, npages);
 
 	if (!cc_platform_has(CC_ATTR_HOST_MEM_ENCRYPT))
-		enc_dec_hypercall(vaddr, npages << PAGE_SHIFT, enc);
+		enc_dec_hypercall(vaddr, (unsigned long)npages << PAGE_SHIFT, enc);
 
 	/*
 	 * On CSV3, the shared and private page attr changes should be managed
